Reorganização da estrutura da tela de perfil portfólio
