+++ src/Screens/ProfilePortfolio/styles.ts
@@ -1,0 +1,37 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+  safeArea: { flex: 1, backgroundColor: '#FFFFFF' },
+  page: { flex: 1, backgroundColor: '#FFFFFF' },
+  scrollContent: { paddingBottom: 32 },
+  backButton: { height: 92, paddingHorizontal: 24, flexDirection: 'row', alignItems: 'center', gap: 6 },
+  backText: { color: '#111111', fontSize: 21, fontWeight: '700' },
+  cover: { height: 188, width: '100%' },
+  profileContent: { paddingHorizontal: 24, position: 'relative' },
+  avatar: { width: 164, height: 164, borderRadius: 82, borderWidth: 6, borderColor: '#FFF', alignSelf: 'center', marginTop: -101 },
+  headingRow: { marginTop: 11 },
+  headingText: { width: '100%' },
+  name: { color: '#0A0A0A', fontSize: 33, fontWeight: '700', letterSpacing: -0.8, flexShrink: 0 },
+  locationRow: { flexDirection: 'row', alignItems: 'center', gap: 7, marginTop: 6 },
+  city: { color: '#7D7D7D', fontSize: 16, fontWeight: '600' },
+  occupation: { color: '#7D7D7D', fontSize: 16, fontWeight: '600', marginTop: 10 },
+  cnpjBadge: { position: 'absolute', top: 14, right: 24, borderWidth: 1.5, borderColor: '#B6B6B6', borderRadius: 20, paddingHorizontal: 8, paddingVertical: 2, flexDirection: 'row', alignItems: 'center', gap: 4, backgroundColor: '#FFFFFF' },
+  star: { color: '#D4AE54', fontSize: 13 },
+  cnpjText: { color: '#7A7A7A', fontSize: 12, fontWeight: '600' },
+  aboutCard: { height: 370, borderWidth: 1.5, borderColor: '#777', borderRadius: 29, marginTop: 42, paddingHorizontal: 40, paddingTop: 17 },
+  aboutTitle: { color: '#0A0A0A', fontSize: 21, fontWeight: '700', marginBottom: 18 },
+  bioScroll: { flex: 1, marginBottom: 18 },
+  bioContent: { paddingRight: 18 },
+  bio: { color: '#777', fontSize: 16, lineHeight: 21, textAlign: 'justify' },
+  workSection: { marginTop: 36 },
+  workTitle: { color: '#0A0A0A', fontSize: 23, fontWeight: '700', marginBottom: 28 },
+  workRows: { gap: 18 },
+  workRow: { gap: 16, paddingRight: 24 },
+  workImage: { height: 172, width: 210, borderRadius: 18, borderWidth: 1, borderColor: '#E4E4E4', shadowColor: '#000', shadowOpacity: 0.18, shadowRadius: 6, shadowOffset: { width: 0, height: 3 }, elevation: 4 },
+  imageModal: { flex: 1, alignItems: 'center', justifyContent: 'center' },
+  modalBackdrop: { ...StyleSheet.absoluteFillObject, backgroundColor: 'rgba(0, 0, 0, 0.86)' },
+  imagePreviewContainer: { width: '92%', height: '72%', position: 'relative' },
+  zoomableImage: { width: '100%', height: '100%' },
+  imagePreview: { width: '100%', height: '100%', borderRadius: 16 },
+  closePreviewButton: { position: 'absolute', top: -12, right: -8, width: 42, height: 42, borderRadius: 21, backgroundColor: '#202020', alignItems: 'center', justifyContent: 'center', borderWidth: 1, borderColor: '#FFFFFF' },
+});
